Fix CFBundleVersion for Carthage Installs (#107)
* Set CURRENT_PROJECT_VERSION and fix Info.plist references

Uploading to the App Store was failing when Seam3 was installed via Carthage. This is because `CURRENT_PROJECT_VERSION` was not set in the `project.pbxproj` file so when it compiled the `Info.plist` file, the `CFBundleVersion` was empty and the App Store was rejecting uploads.

There were 2 options for a fix here:
1. Update the `CURRENT_PROJECT_VERSION` in each of the targets and the respective Info.plist files would pick that up. This is kind of hidden since you have to go into the Target Build Settings and find the setting.
2. Update the Info.plist files directly. This felt more straight-forward since the Info.plist files are right at the top level of the project and easy to find.

I set the `CFBundleVersion` in each of the `Info.plist` files to the current version.

I also organized the macOS and tvOS Info.plist files. The references were broken in the `pbxproj` file and the iOS/watchOS Info.plist files were organized in the `Assets/Resources` directory so now they're all together.

* Update to static CURRENT_PROJECT_VERSION