Fix multithreading MOC issue (#108)
`SMStore.resetBackingStore` gets called from the host app so it will be on a different thread than the `backingMOC`. This adds a `performAndWait` before executing a request on the `backingMOC` so it's run on the same thread.

You can't throw an exception from inside the `performAndWait` since it's a callback so I added an `Error?` variable outside of the context to capture the error if one is thrown. This retains the original behavior.